LIGHTWEIGHT_UI: Allow feedrate % to exceed 255% (#11656)
- Changed type of integer used for feedrate percentage from uint8_t to uint16_t.

+++ Marlin/src/lcd/dogm/status_screen_lite_ST7920_class.h
@@ -87,7 +87,7 @@ class ST7920_Lite_Status_Screen {
     static void draw_bed_temp(const int16_t temp, const int16_t target, bool forceUpdate = false);
     static void draw_fan_speed(const uint8_t value);
     static void draw_print_time(const duration_t &elapsed);
-    static void draw_feedrate_percentage(const uint8_t percentage);
+    static void draw_feedrate_percentage(const uint16_t percentage);
     static void draw_status_message(const char *str);
     static void draw_position(const float x, const float y, const float z, bool position_known = true);
